requirements.txt
numpy==1.21.6
pygame==2.6.0
opencv-python==4.10.0
networkx==2.6.3
matplotlib==3.5.3
gym==0.26.2
tqdm==4.66.5
PyYAML==6.0.1
pandas==1.3.5